# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: VendorIQ Procurement Intelligence API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /contracts:
    get:
      operationId: listContracts
      tags: [contracts]
      summary: List analyzed contracts
      parameters:
        - in: query
          name: supplierId
          schema:
            type: ["string", "null"]
        - in: query
          name: status
          schema:
            type: ["string", "null"]
            enum: [active, expiring, expired, under_review, null]
      responses:
        "200":
          description: Contract list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contract"

    post:
      operationId: analyzeContract
      tags: [contracts]
      summary: Upload and analyze a contract
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContractInput"
      responses:
        "201":
          description: Analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContractDetail"
  /contracts/{id}:
    get:
      operationId: getContract
      tags: [contracts]
      summary: Get full contract analysis
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Contract detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContractDetail"
  /contracts/{id}/clauses:
    get:
      operationId: getContractClauses
      tags: [contracts]
      summary: Extracted and classified clauses
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Clause list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContractClause"

  # ─── Knowledge Graph ──────────────────────────────────────
...
  /alerts:
    get:
      operationId: listAlerts
      tags: [alerts]
      summary: Smart alerts queue
      parameters:
        - in: query
          name: read
          schema:
            type: ["boolean", "null"]
        - in: query
          name: type
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Alert list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SmartAlert"
  /alerts/{id}/read:
    post:
      operationId: markAlertRead
      tags: [alerts]
      summary: Mark alert as read
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Updated alert
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SmartAlert"
  /alerts/{id}/dismiss:
    post:
      operationId: dismissAlert
      tags: [alerts]
      summary: Dismiss alert
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Dismissed

  # ─── Agents ───────────────────────────────────────────────
...
  /agents/{agentId}/run:
    post:
      operationId: runAgent
      tags: [agents]
      summary: Trigger an agent to run its next task cycle
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Updated agent status after triggering run
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentStatus"
        "404":
          description: Agent not found
  /agents/query:
    post:
      operationId: queryAgent
      tags: [agents]
      summary: Send natural-language query to Procurement Copilot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentQuery"
      responses:
        "200":
          description: Agent response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentResponse"

# ─── Components ─────────────────────────────────────────────
components:
  schemas:
...
    Contract:
      type: object
      required: [id, supplierId, supplierName, title, status, value, expiresAt, riskScore]
      properties:
        id:
          type: string
        supplierId:
          type: string
        supplierName:
          type: string
        title:
          type: string
        status:
          type: string
          enum: [active, expiring, expired, under_review]
        value:
          type: number
        expiresAt:
          type: string
        riskScore:
          type: ["number", "null"]
        clauseCount:
          type: ["integer", "null"]
        riskyClauseCount:
          type: ["integer", "null"]
    ContractInput:
      type: object
      required: [supplierId, title, content]
      properties:
        supplierId:
          type: string
        title:
          type: string
        content:
          type: string
          description: Full contract text or base64-encoded PDF
        value:
          type: ["number", "null"]
        expiresAt:
          type: ["string", "null"]
    ContractDetail:
      type: object
      required: [id, supplierId, supplierName, title, status, value, expiresAt, summary, riskScore, missingClauses, complianceIssues]
      properties:
        id:
          type: string
        supplierId:
          type: string
        supplierName:
          type: string
        title:
          type: string
        status:
          type: string
        value:
          type: number
        expiresAt:
          type: string
        summary:
          type: string
        riskScore:
          type: number
        missingClauses:
          type: array
          items:
            type: string
        complianceIssues:
          type: array
          items:
            type: string
    ContractClause:
      type: object
      required: [id, clauseType, text, riskLevel, explanation]
      properties:
        id:
          type: string
        clauseType:
          type: string
        text:
          type: string
        riskLevel:
          type: string
          enum: [low, medium, high]
        explanation:
          type: string
        recommendation:
          type: ["string", "null"]

    # ── Knowledge Graph ──
...
    SmartAlert:
      type: object
      required: [id, type, title, message, severity, isRead, createdAt]
      properties:
        id:
          type: string
        type:
          type: string
          enum:
            - high_risk_supplier
            - contract_expiring
            - price_spike
            - drift_detected
            - pipeline_failure
            - negative_news
            - fraud_detected
        title:
          type: string
        message:
          type: string
        severity:
          type: string
          enum: [info, warning, error, critical]
        isRead:
          type: boolean
        supplierId:
          type: ["string", "null"]
        supplierName:
          type: ["string", "null"]
        createdAt:
          type: string
        dismissedAt:
          type: ["string", "null"]

    # ── Agents ──
    AgentStatus:
      type: object
      required: [id, name, type, status, lastRun, tasksCompleted]
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
          enum: [risk_analyst, price_forecaster, contract_analyst, fraud_investigator, procurement_copilot]
        status:
          type: string
          enum: [idle, running, error]
        lastRun:
          type: ["string", "null"]
        tasksCompleted:
          type: integer
        currentTask:
          type: ["string", "null"]
        avgResponseMs:
          type: ["number", "null"]
    AgentQuery:
      type: object
      required: [message]
      properties:
        message:
          type: string
        agentType:
          type: ["string", "null"]
          enum: [risk_analyst, price_forecaster, contract_analyst, fraud_investigator, procurement_copilot, null]
        context:
          type: object
          additionalProperties: true
    AgentResponse:
      type: object
      required: [agentType, message, confidence]
      properties:
        agentType:
          type: string
        message:
          type: string
        confidence:
          type: number
        sources:
          type: array
          items:
            type: string
        actions:
          type: array
          items:
            $ref: "#/components/schemas/AgentAction"
    AgentAction:
      type: object
      required: [label, type]
      properties:
        label:
          type: string
        type:
          type: string
        payload:
          type: object
          additionalProperties: true