# lib/api-zod/src/generated/types/agentQuery.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * VendorIQ Procurement Intelligence API
 * OpenAPI spec version: 0.2.0
 */
import type { AgentQueryAgentType } from './agentQueryAgentType';
import type { AgentQueryContext } from './agentQueryContext';

export interface AgentQuery {
  message: string;
  /** @nullable */
  agentType?: AgentQueryAgentType;
  context?: AgentQueryContext;
}