# lib/api-zod/src/generated/types/smartAlert.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * VendorIQ Procurement Intelligence API
 * OpenAPI spec version: 0.2.0
 */
import type { SmartAlertSeverity } from './smartAlertSeverity';
import type { SmartAlertType } from './smartAlertType';

export interface SmartAlert {
  id: string;
  type: SmartAlertType;
  title: string;
  message: string;
  severity: SmartAlertSeverity;
  isRead: boolean;
  /** @nullable */
  supplierId?: string | null;
  /** @nullable */
  supplierName?: string | null;
  createdAt: string;
  /** @nullable */
  dismissedAt?: string | null;
}